Fixes for var precedence
Don't require CCLD to be defined in settings.conf
If not, use the value of $CC
It is C_FLAGS, LD_FLAGS, etc. which should be cleared per-subdir, not CFLAGS, etc.
Ensure that command line vars take precedence over everything
Forget to add local uninstall target (is this correct?)
test2 needs to UseSystemLibs $LIBS
Also fix state handling when reloading

Signed-off-by: Steve Bennett <steveb@workware.net.au>