Improvements to dyndep
When scanning for dyndeps, if a target fails to build, still included it as a dyndep
When scanning for dyndeps, include both the first file found and the file in "." if it exists.
This is to account for #include "file.h" vs #include <file.h> where we don't know which one
might be selected.
Also add --quiet to suppress timing message