allow rules to be redefined
If a rule is redefined identically, apart from source location,
allow it with a warning. This makes it possible to do, e.g.

Executable a a.c x.c
Executable b b.c x.c

Otherwise the Object rule for x.c would be invoked twice, giving an error.
This needs to be rethought, but will do for now