dynamic dependency scanning - cache was overly aggressive
Consider a.c including b.h and the result being cached.
Now if b.h is modified to include c.h, the scan cache was still
considered up-to-date, even though the scan should be rerun to pick
up the dependency on c.h

This is now fixes by including the mtime of each result in the cache.
If any mtime changes, a rescan is forced.

Also, the remove the (incorrect) code to discard a orphan target
during header scanning. It isn't needed.