Always rebuild targets with -do and no cached info
Even if no dependencies. This means that targets that are wholly
defined by their -do script will still be re-run if the script changes.
(See make-config-header in uweb.git)

There may be cases where we don't want this behaviour, but if necessary
we can add a flag to select that.