don't exit early due to error if in --configure
It is possible to have a rule like:

Install --bin $bindir file

And due to some error (say in auto.def), $bindir is not defined. Now tmake fails to run anything, including ./configure
making it hard to recover from this error.
In this case it is best to ignore the errors and rebuild
configure if possible.

More nuance may be required here